move scanning snapshotter out of tests and clean up list backup command
