+++ src/test/kotlin/com/salesforce/dynamodbv2/mt/mappers/MtBackupManagerTest.kt
@@ -68,7 +68,7 @@ internal class MtBackupManagerTest {
         val sharedTableBinaryHashKey = SharedTableBuilder.builder()
                 .withAmazonDynamoDb(dynamo)
                 .withContext(mtContext)
-                .withBackupSupport(s3Client, "fake-bucket", MtSharedTableBackupManagerS3It.MtScanningSnapshotter())
+                .withBackupSupport(s3Client, "fake-bucket", MtScanningSnapshotter())
                 .withTruncateOnDeleteTable(true)
                 .withBinaryHashKey(true)
                 .build()
@@ -100,7 +100,7 @@ internal class MtBackupManagerTest {
         val sharedTableBinaryHashKey = SharedTableBuilder.builder()
                 .withAmazonDynamoDb(dynamo)
                 .withContext(mtContext)
-                .withBackupSupport(s3Client, "fake-bucket", MtSharedTableBackupManagerS3It.MtScanningSnapshotter())
+                .withBackupSupport(s3Client, "fake-bucket", MtScanningSnapshotter())
                 .withTruncateOnDeleteTable(true)
                 .withBinaryHashKey(true)
                 .build()
